requirements.txt
scikit-learn
pillow
torch
facenet-pytorch
streamlit


